+++ pkg/azuremetadataclient/azuremetadataclient.go
@@ -1,4 +1,4 @@
-package azuremetadata
+package azuremetadataclient
 
 import (
 	"bytes"
