+++ pkg/eventhandler/interface.go
@@ -3,9 +3,9 @@ package eventhandler
 import (
 	"context"
 
-	"github.com/giantswarm/azure-scheduled-events/pkg/azuremetadata"
+	"github.com/giantswarm/azure-scheduled-events/pkg/azuremetadataclient"
 )
 
 type EventHandler interface {
-	HandleEvent(ctx context.Context, event azuremetadata.ScheduledEvent) error
+	HandleEvent(ctx context.Context, event azuremetadataclient.ScheduledEvent) error
 }
